requirements.txt
django~=5.2.5
channels~=4.3.1
django-filter~=25.1
asgiref~=3.9.1
whitenoise~=6.11.0
python-decouple
daphne